Introduce Keep Vendor contract
Introduced Keep Vendor contract according to Keep RFC 12.

A vendor is a smart contract with which the application interacts to
obtain the new instance of a keep of the given type backed by enough
stakers so that all operations on the created keep can go smoothly.
Vendor interacts with keep factories choosing the one most recent if
it is backed by enough stakers. Each vendor is a service contract
and does not have to be sanctioned by stakers. Each vendor can
operate on several versions of keep factory to allow for a smooth
upgrade process.