Provide script for migrating contracts via CircleCI
This is largely a copy/paste job from `keep-core`.  Here we provide the
script that gets run on our utility box inside a target VPC to migrate
contracts for a given project.  This is to be used in conjunction with
CircleCI and a job config that provides the required configs.  Required
configs are largely sourced from the CircleCI context we use along with
the job configuration.  Default is our private testnet `keep-dev`.