Remove truffle_sample -> truffle file copy
In keep-tecdsa the truffle file in source control is named truffle.js,
not truffle_sample.js.  This caused the removed `cp` command to fail on
file not found.  I'm removing this command since the file is named as
expected by truffle.  We should probably keep a proper copy of
`truffle.js` in `keep-core` for the sake of consistency.