Change keep_dev truffle host to localhost
For Circle run migrations we use kubectl port-forward to expose the
eth-tx node to the utility box.  We do this because the utility box is
outside of the Kube cluster and we don't expose the eth-tx node to the
internet on the private testnet.  Since we use kubectl port-forward
we're connecting to the localhost and not the host directly.