Build and publish initcontainer-provision-keep-tecdsa
Like the keep-client, keep-tecdsa depends on deployed ETH contracts and
ETH accounts.  Development environments have ever changing contract
addresses which we need to account for in provisioning of keep-tecdsa.
To manage this dynamically we configure an Initcontainer to setup a
configuration file with the latest contract information and a fresh eth
account.  For now we're not using the ETH account as there's a hardcoded
solution for a demo.