Extract contract's package.json to root directory
We moved package.json for solidity contracts to a root directory. It's a
workaround for npm limitation that doesn't allow to install dependencies
from git repositories from a subdirectory.
With this change we will be able to install contracts via npm in
repository which needs them.
See: https://github.com/npm/npm/issues/2974