Find recovery ID (v) for signature
We evaluate `v` value for a signature based on the solution in bitcoin.
V is a recovery ID which helps to extract public key from a signature.
A signature can be linked to up to 4 points representing a public keys.
With V value we can determine which point was the same as signer's
public key.