Replace Membership struct with Signer
We don't need Membership structure to hold keep and Signer as we store
signers in a map where keep address is a key and Signer is value. So it
seems like unnecessarry duplication. Here we remove Membership structure
and related code and refactor storage functions to handle separate keep
address (file directory) and signer (file content).