Tag docker images before publication
- We moved tagging a docker image to the publication job. We don't want to
mix contexts between steps. Tagging is specific with variables from
`keep-dev` context and it's strictly related to images publication.
- We also separated docker images building to two steps: first we
build image with target specific to go build environment and run
go tests in this container; later we build a runtime image which
will be published as keep-tecdsa image. Runtime container is a clean
container containing just a client binary.