Convert keep-tecdsa from Deployment to StatefulSet
Here we transition the keep-tecdsa deployment to a StatefulSet.  The
motivation here is that we want a properly persisted DataDir for group
data.  This is required for a client to recover group information if the
container gets cycled.  We're setting the DataDir path via the env var
KEEP_DATA_DIR in the InitContainer env: section.  The InitContainer will
write this path to the config section Storage.DataDir.  We need to mount
a persistent volume to this path, which we're doing.