Testing git tag filters and manual approval
Rather than maintain a separate branch for keep-test deployment we want
to properly integrate with Circle.  For now this requires two things in
addition to the new "by environment" workflows.  First, we need to
initiate a keep-test deployment only when a new git tag/release is cut.
The tag filter acomplishes this.  Second, we need to initiate the
migration and subsequent builds only when we're ready to deploy not
immediately when a tag is cut.  To handle this we introduce manual
approval to kick off the keep-test workflow.