Added net interfaces to cover unicast channels
Added provider interface which should be used to obtain a network
provider which can provide broadcast and unicast channels for
communication between nodes.
Defined unicast channel interface to implement point-to-point
communication.
Added some aliases to the code from keep-core which are helpful until we
integrate this code into keep-core repo.