Added Group ID to network bridge broadcast channel
We aded Group ID parameters for network bridge to generate globally
unique name for broadcast channel. It can happen that the same members
are selected to different groups but with previous solutions they all
would use the same broadcast channel. Now we need to provide unique name
for each group.