Add option for not pooling tickets at all
This makes me feel a little bit stupid for not realizing earlier
but calculating the number of tickets we request
after we get the selection seed from the beacon
means that we get pooling "for free" with high levels of demand
with very little implementation complexity.

This would un-bottleneck us on both beacon throughput and development time.