Temporary change to read config file from ConfigMap
The code deployed to keep-test is old enough that it doesn't have my
recent changes to make keyfile path configurable on via the
provisioner InitContainer.  This results in keep-test keep-tecdsa still
getting a fresh ETH account on cycle.  Here we bypass the InitContainer
all together and read a complete config file from ConfigMap.  This
should be undone when we deploy more recent code to keep-test.