Script to provide bonding value
A script to provide a bonding value to a KeepBonding contract. The script requires contract to be provided in web3 Contract type so we can use the script in both local development environment (with Truffle) and in CI deployment provisioning where we use strictly web3.
The script funds a bonding value from specific account for given operator.
If the available bonding value is greater or equal required bonding value it does nothing.