Changed MemberID to uint32 type
We no longer require MemberID to be a bytes representation of ethereum
address, so we simplify the type behind member ID to be an uint32. This
is expected to correspond with member index in the keep signing group.
Please note that due to TSS protocol requirements indexing should start
with 1 - member ID equal to zero is not allowed.