Added attach_workspace to migrate_contracts job
For `migrate_contracts` we require `setup_github_package_registry` job
to be executed. The job stores github token in .npmrc file which we
persiste to a workspace. With this change we will attach the workspace
to the migrations job so packages can be imported from private github's
packages for @keep-network scope.