Removed openKeep from ECDSA Keep Vendor
We removed openKeep function from ECDSA Keep Vendor contract as we want
the function to be called directly on factory contract. The factory will
want to verify if the caller of the keep opening function is eligible
application and doing it through vendor contract prevents it of doing
so.
We expect that external application will call vendor contract to get a
recommended factory address and then the application will call the
given factory to open a new keep.