One client run multiple operators
We implemented changes so one client is able to run multiple operators.
This is a temporary solution which should be completely reverted as soon
as proper networking (unicast channel) is implemented. We expect client
config file to contain paths to multiple ethereum key files. The client
will initialize for each configured key file.