Temporary changes to deployment to support multi operator client
We introduceed temporary changes to deployment scripts to support
running multiple operators on single client instance.
The changes here are a temporary solution and should be completely
reverted as soon as we've got proper networking implemented (unicast).
It is required that config.toml contains a list of paths to three unique
ethereum key files, here we generate the files update configuration of
kubernetes stateful set.