Pass ether from member registration to bonding
Operator calling `registerMemberCandidate` function can transfer a value
with the call. The value will be forwarded to keep bonding contract with
`deposit` function call to register the amount as available for bonding.
Besides this way an operator will still have a possibility to call the
bonding contract directly to deposit a value for bonding.