Added destination paratemer to seizeBond function
Caller of the function (who is expected to be the holder of the bond)
will have a possibility to transfer the bond to a specific destination.
It not necesarrily need to be transfered to the holder.
This will be used by ECDSA keep to transfer bonds to the owner of the
keep.