Call random beacon to request a new relay entry
After using group selection seed to select new group we need to request
a new seed value for the future usage. We call teh random beacon
asynchronous function which when entry is ready will set it for as in a
callback. Until this is done we have to recalculate the seed to a new
value so we simply hash it to obtain a new value.