Validate malleability of s in submitSignature function
As per EIP-2 and BIP-62 to prevent malleability it is required that `s` part of the
signature is a number in the lower half of the secp256k1 curve order.
The problem with ECDSA sigantures is that for the given digest it is
possible to succesfully verify two signatures: `(r, s)` and `(r, -s)`.
Most of the ECDSA implementations produce and accepts signatures with
`s < N / 2`, where N is an EC order.

We require that signature provided to the `submitSignature` function has
`s` value lower or equal than the half of the secp256k1 curve's order.

If the value calculated by the signer is in the high half, it is
signer's responsibility to calculate `s' = N - s`. It also requires
updating `recoveryID` by changing it's value: `0 <-> 1` or `2 <-> 3`.