Accumulate funds sent by the beacon in the subsidy pool
Random beacon sends back callback surplus to the requestor. It may also
decide to send additional request subsidy fee. What's more, it may
happen that the beacon is busy and we'll not refresh group selection
seed. We accumulate all funds received from the beacon in the subsidy
pool of the factory and later distribute funds from this pool to
selected signers.