Added closeKeep function
Function can be used to close the keep. Closing the keep sets if state
to inactive by the `isActive` flag. It releases bonds of all members and
emits an event.
Only owner of the keep can call the function.
Function can be called only if keep is in active state, so can't be
called twice.
Keep can be closed only if there is no signing process currently going
on or if the current signing process times out.