Hardcoded gas limit for submit public key function
All members try to submit public key at once and it is impossible to get
an accurate estimation of a transaction. We are setting the gas limit to
make it enough for a group of 3 members in any case but we'll have to
revisit this approach soon in the future.