Set ethereum node address from ETH_URL env variable
We added setting ethereum address from ETH_URL env variable.
The address in the config is not always of form `HOSTNAME:PORT` we added
possibility to provide the address as an URL.
We may want to update all our configs to use just URL form, but since we
do that we gave an alternative possibility to set it the old way.