Always create a sortition pool with a minimum bond of 1
Sortition pool is created with a minimum bond of 1 to avoid griefing.

Anyone can create a sortition pool for an application.
If a pool is created with a ridiculosly high bond, nobody can join it
and updating bond is not possible because trying to select a group with
an empty pool reverts.

We set the minimum bond value to 1 to prevent from this situation and to
allow the pool adjust the minimum stake during the first signer
selection.