Moved sortition pool creation to lcl-initialize script
Sortition pool is scrictly related to the dev environment configuration
as it uses customer application address. Also in the production
environment we will first need to deploy contract and after that when
TBTCSystem is deployed we will get an address of the tbtc contract for
which client should be initialized.