Limit maximum group size to 16 signers
To not letting one malicious group member decide on group's public key
we decided all group members need to publish generated public key and
only if all the keys matches we accept the result as keep's public key.

Since all members submits the key at once at the end of key-gen
protocol, estimating exact transaction cost is not possible as it
depends on the order in which submitPublicKey transactions will be
picked by miners. For that reason, we need to set the gas limit for this
transaction to the most pessimistic case.

The most pessimistic case gas cost depends on the group size. We've run
a number of experiments and for a group size of 16 members, gas limit
set to 300k is enough.

Gas limit on submitPublicKey transaction is not the only reason why it
makes sense to limit the max group size. Each signing group member node
will need to have a direct connection with all other group members. It
impacts node discovery and network topology. Allowing to pass an
arbitrary high number may hurt the network.

It's safer to allow for something we are sure about is safe and works
rather than allow for any value and cry later.