Do not revert in availableUnbondedValue
Sortition pools check the availableUnbondedValue and if the operator is
not up to date it's skipped and removed from the pool. If we revert in
this function, we'll fail operator selection. The current one and all
future ones as the sortition pool will not be able to remove outdated
operator until someone updates it.