Wait until the operator is registered in a pool
Added a new function that waits until the operator is registered.
Previously, we had a situation that for every new registered operator we
were triggering the monitoring right away. This was problematic because
all monitoring checks were instantly failing as the transaction was
usually not yet mined. From now on, before we start monitoring, we wait
until the operator is confirmed as registered in the pool.