Bring back write lock to UnregisterKeep, eliminate ForEachKeep
Write lock in UnregisterKeep makes sense because we are modifying the
state.

ForEachKeep function is problematic and leads to deadlocks so we
eliminated it and use a combination of GetKeepsAddresses + GetSigners.

ForEachKeep uses a read lock. If any function inside the callback wants
to use a write lock, we'll have a problem. Another option here was to
snapshot the state inside of ForEachKeep and iterate on the snapshot
without locking but the current solution seems less error-prone.