Two-step update of vendor implementation address
Here we introduce two-step update process for vendor implementation
address. To update the value it is first required to use `upgradeTo`
function to provide a new value and emit an event that upgrade has
started. After defined time delay period it is required to call
`completeUpgrade` function to actually change the value.