Seize signer bonds only when active
Added requirement to seizeSignerBonds function to check if the keep is
still active when calling the function.
When the function is called we want to close the keep so we need to
check if it hasn't been already closed. When keep is closed we either
seize the bonds or free them.