Code requesting relay entry extracted, removed gas limit
Call gas limit makes sense when an operator contract calls
a service contract and the transaction failure may result
in slashing. In this case, it's the customer who will decide
whether the estimated gas amount is reasonable. And if the
transaction fails, new keep will not get open so no one can
be slashed. With no gas limit we are safer in case of opcode
price change.