Removed notifyKeepClosed
BondedECDSAKeepFactory is authorized for slashing so in case of
a signature fraud, keeps ask the factory to slash members. The
factory needs to validate if the keep asking for slashing is still
active and if it was created by that factory. Closed keeps,
terminated keeps, and keeps not created by the given factory can
not ask for slashing.

So far, this was all done by onlyKeep modifier in the factory.
notifyKeepClosed executed on the factory was making the given keep
instance no longer eligible for asking for slashing.

Since now we need to track all keeps the factory ever opened, this
commit refactors onlyKeep to use creationTime to check if the keep
was created by the factory and later, confirm if the keep is still
open checking its status. This way, we can get rid of
notifyKeepClosed.