Remove ETH account files
We now fetch ETH account info from ConfigMaps. The files here were a bit
of a band-aid to keep-tecdsa running while we were on the "one client
multiple accounts" deployment.  Now that we have networking for
keep-tecdsa we can split the deployments up and these static files are
no longer needed.