Set initialized flag in implementation contructor
We want to set the flag that implementation contract has been already
initialized in the constructor to block direct usage of the contract.
The contract is intended to be used via proxy only, we don't want to
cause any confusions for users calling the contract directly. This
change also blocks possibility of calling initialize function on
deployed contract, which could be a possible attack vector.