Added upgradable proxy storage contract
This contract can be used to hold implementation details in a two-step
upgradeable proxy to hold details of the implementation. In proxy pattern
data should be stored in a ways that reduces possibility of collisions between
proxy and implementation contracts. In this contract we use a mapping which
is allocating storage slot for the data in a dynamic way.