Moved implementation initialization to second upgrade step
Previously we initialized implementation contract in the first step of
the upgrade that gave some possibility to interactions with the new
contract before it should actually be used.

Here we move the initialization call to complete upgrade function and
update handling of the implementation details to use the ones from
UpgradableProxyStorage.