Moved implementation initialization to second upgrade step
Previously we initialized implementation contract in the first step of
the upgrade that gave some possibility to interactions with the new
contract before it should actually be used.

Here we move the initialization call to complete upgrade function and
update handling of the initialization details to be able to store in the
first step and use in the second step.

We use mapping as it's allocating storage based on keccak256 hash of the
entry so it lowers a possibility of clashing with implementationd
fields.