Scale out ecdsa deployment
We're about to start a round of system tests that require a larger
network than what we run by default.  To support this we've added 5 more
ecdsa deployments to the keep-test environment.  It's worth nothing that
the current approach for adding a new client (statefulset + service
config file) wont scale too well.  We're going to look at collapsing the
required config bits down into a single template file we can iterate on
for deployment.