Generate PreParams with provided optional concurrency
tss-lib library has optional paramter of concurrency, if the value is
not provided it will get number of runtime CPU. On machines with CPU
number less that 3 it fails later due to the calculations the function
is performing where the number is divided by 3. It is a bug described in
https://github.com/binance-chain/tss-lib/issues/93

As a workaround we provide a fixed value of 3 so the calculation will
give valid results.