Add a concept of terminated keep, track timestamps in factory
To properly implement stakedrop rewards we need to track the creation
timestamp for each keep as well as know which keeps have been properly
closed. This is achieved in this commit.

Instead of using a simple isActive boolean allowing to represent only
two states, we have an enum allowing to indicate the keep is either
active, closed, or terminated. Keeps is getting terminated when bonds
are seized.

The factory keeps information about keeps it created as well as their
timestamps. This allows to extract all the information needed by
stakedrop rewards contract. Also, we no longer have to notify the
factory when a keep is closed.