Configurable TSS pre-param generation concurrency level
We add a configurable parameter to let operator define concurrency level
for pre-params generation in TSS protocol.

The level should be configured based on the resources available on the
machine where the client is run.

Due to a bug in tss-lib it cannot be lower than 3, we enforce it by
updating provided value to be no less than 3.

The parameter is optional, if not provided `3` will be set.