Prepare Solidity package.json for auto-publishing
Auto-publishing is already happening on keep-core, with a
build-time-decided -pre build going out on every master merge. The
package.json file for keep-ecdsa now references the latest version of
that -pre build, and is ready to start including artifacts with the
keep-ecdsa build as well. Finally, the version for keep-ecdsa is aligned
to the keep-core version.