Updated TSS protocol to reuse pre parameters between retries
Previously when signer generation failed there was retry mechanism
invoking the process once again. Unfortunately it wasn't efficient as
pre-parameters had to be generated again, even though they were not used
for actual protocol execution, as the failure could happen before the
messages were exchanged between parties.

Here we introduce a change that will reuse pre parameters for protocol
execution up to the moment when the messages are actualy exchanged
between parties. Before running the phase where parties interact we
flush the parameters so if the phase fails we will need to provide fresh
set of parameters.