Added structs to track requests from events
It may happen that we receive the same event multiple times. We don't
want to kick off duplicated processes for such events. We introduced a
way to track requests and start only single processes for them.

We register event in the struct after receiving it. If it's already
there we will log an error an return from the process. If it's unique
event we will start the process. When we return from the process we will
remove the event from the tracking struct.