Wait for signature request chain confirmation on start
When the client receives signature request from chain in an event it
waits for the request to be confirmed. Here we add the same logic for
signing requests picked after client start. When client starts and sees
that some keep is watiing for their signature it will wait for specific
blocks and check that the signature is really requested.