Minimum stake schedule
Minimum stake is no longer a system-wide constant and it changes over
time in TokenStain. In this commit, BondedECDSAKeepFactory is updated to
support minimim stake schedule.

The factory does no longer have the minimum stake hardcoded. To check if
the operator has minimum stake, it queries TokenStaking contract.

Also, the sortition pool now expects the recent value of minimum stake
to be passed for each group selection. This is what is happening in
openKeep function.

Factory declares the eligible stake divisor for sortition pool to be 1
KEEP.