Trigger operator updates by events
Here we updated trigger for operator status update. We no longer check
if operator is up to date on each block but listen for events that may
have affected operator's available bonding value or token stake that is
the parameters that are verified by sortition pool to check operator
eligibility.