Individual context for public key submission monitoring
Previously the context for public key submission monitoring was a child
of public key generation context. Once the operator published their key
the parent context was completed so the monitoring routine was
cancelled. Here we define a seperate unique context for monitoring
function with other timeout.

There is abort channel so the monitoring can be aborted in case of key
submission failure. If everytihing goes well the monitoring will track
chain after the signer establishement.