Time cache implementation for active keeps firewall rule
We will have to keep the list of active keep members in a cache for a
certain duration. Figuring out if the given client is a member in at
least one active keep will take a bunch of calls to Ethereum client and
to minimize them for every single firewall check tick, we can cache the
result in-memory.

This will also allow us to share check results between multiple
connections.