Minimum stake or active keep membership policy
For ECDSA keep client, we always connect with another client if it has
minimum stake. In a situation when the remote client has no minimum
stake we should check if it is a member in at least one active keep. If
so, we allow to establish the connection.

The reason for that is to allow the client to participate in any
oustanding signing in case it's been slashed, most probably by the
beacon. The client with no minimum stake is not eligible to be selected
to new ECDSA keeps but it can still participate in signing in existing
keeps it's a member of.