Fixed logging of signature submission error
If there was an error for `SubmitSignature` we were checking
`IsAwaitingSignature` and overwritting the error received from the
`SubmitSignature` function call. If function execution got up to `failed
to submit signature for keep` log the error logged was `nil` as it was
replaced by `IsAwaitingSignature` error.
Here we update the variable name so it is unique and won't be
overwritten.