hasAuthorization function on BondedECDSAKeepFactory
The function allows to check if the factory has the authorization to
operate on stake represented by the provided operator.

We will need this function to optimize the number of Ethereum client
calls for firewall rule when the operator has no minimum stake.

When the remote peer has no minimum stake, the firewall goes through all
keeps in the factory trying to find at least one active keep the given
peer is a member of to decide whether the connection can be open or not.

Although results are cached, for every remote peer that has never been a
member of any keep, we'll keep iterating through those keeps trying to
find the one matching and attacker can quickly drain the maximum allowed
number of requests to Ethereum client if there is any limit.

To fix this, we should check if the operator has authorization on the
factory. If the operator has no authorization it means it could never
participate in any keeps. This check will be added in a separate commit
to the firewall.