Firewall checks if the remote peer has authorization on the factory
The authorization is permanent and can not be revoked.
If the peer has no authorization on the factory it means it could never
participate in signer selection and there is no chance it is a member of
any keep.

We do it to avoid draining Eth call limit by malicious members with no
minimum stake who has never participated in the network. Without
authorization check, we'd have to go through all keeps all over again
for each such connection attempt.