Start signature confirmation from block it actualy was requested
Previously we started confirming the signature request on client start
from the current block. But the signature may have been requested couple
of blocks before so the client was waiting too long compared with other
peer clients.

With this change we start confirmation from the block the signature was
actually requested.