Check if signing has timed out on client start
When client gets started it checks if there is a digest that need to be
signed for the keep. First it waits for 12 blocks to confirm the chain
state. It may happen that the request timeouts in the meantime or it is
timed out right from the begining. Here we confirm that there is no
signing timeout so we don't start the signing process needlessly.