Check if signing timed out on chain in publication function
We check if the signing already timed out on-chain before submitting the
signature. It won't make sense to publish a signature in such case and
execute subsequent steps in case signature gets rejected.

It also covers us when we get passed this check over to the submission part and it fails
because of a timeout happening in a tiny window between
`HasSigningTimedOut` and `SubmitSignature`. In such scenario retry will
happen and we will get the timeout on the next loop run.