Do not revert signature fraud submission TX if slashing failed
In practice, the situation when KEEP token slashing failed should never
happen. However, we want to be very explicit in this contract and
protect the owner that even if it happens, the transaction submitting
fraud proof is not going to fail and keep owner can seize and liquidate
bonds in the same transaction.