Increased firewall active keep cache lifetime to 7 days
If there is an active keep with no minimum stake in the network, in the
worst case, we'll have to go through all the keeps created by the
factory every 7 days when establishing connection with that peer for the
first time. Since closing keeps is not very frequent action, 7 days
looks like a better option than doing it every 24 hours. We can save
some ETH client interactions this way.