Change inconsistent cache log level from Panicf to Errorf
Panicf is killing the client and we don't want to kill client on
production because the inconsistent cache state - even if it happens and
it should never happen - does not stop the client from opening keeps and
providing signatures. Ideally, we should have a log level between ERROR
and PANIC (like CRITICAL) but it is not supported by ipfs logging
library.