Use the first 20 bytes of member ID in the file name
After the update to the most recent version of keep-common, file names
longer than 128 characters are no longer supported. We take just the
first 20 bytes of member ID to append it to the file name.

Actually, we could just drop this piece altogether because it is the
same for all file names created by the given client. Leaving it like
that for now.