Do not attempt to generate key if it is available in the registry
We recently had an interesting situation on mainnet when all three
clients in the signing group were restarted before public key submission
transactions were mined. Clients started key generation again, submitted
new keys to the chain (those transactions were later reverted) and
replaced key material stored on disk.

This situation is very unlikely to happen but it happened.

Here we update the code to never start key generation again if the key
material is available in the registry. There are two scenarios possible:
- public key submission transactions are still mining,
- conflicting public key has been submitted.
In both cases, the client should not attempt to generate the key again.

Also, worth mentioning that the other changes to keep client we have in
the queue are making public key submission process faster by
automatically increasing gas price for transaction if it has not been
mined in reasonable time.