Updated keep-core dependency to v1.2.0-rc.1
Updated keep-core dependency to tagged version. The code is the same as
in the previously referenced commit hash but since we are preparing for
a new RC we are replacing commit hash with appropriate version.

keep-core update entails keep-common update to a tagged version, no code
changes here as as well compared to the previously referenced commit
hash.