Added retry delay to registration loops
We added retry delay to loops where it's possible to continue the loop
execution in case of an error. The solution is similiar to the one we added
in #409. If loop execution gets an error for any reason, e.g. problems
with connection to the ethereum API the client ended up in a crazy loop
logging huge amount of errors. This causes disk space drain by the log
file.