Check if beneficiary is set on bond withdraw
It is possible that unbonded value is deposited for the operator before
beneficiary is set for that operator (via stake delegation). In such
case beneficiaryOf will return zero address and the funds will be
transfered to this address. Here we check if the beneficiary is set, if
not the transaction will revert.