Updated holding references to staking contract
As TokenStaking contract has been refactored in keep-network/keep-core#1867
we can use Authorizations and StakeDelegatable contracts calls as these
contracts define the functions we use in bonding. In this commit we use
separately Authorizations and StakeDelegatable instead of TokenStaking
to be more flexible and able to use common code from AbstractBonding for
both KEEP token and ETH bonding.