Updated client not to use contract signing timeout
The contract no longer tracks a timeout for signing so we had to update
the client. We used the HasSigningTimedOut function to determine if
there is a point in signature publication or confirm if the contract is
still waiting for signature. We replaced it with check if the keep is
active. If keep is not active it means it's either closed or terminated
so the bond may have been seized, there is no point in submitting a
signature in such situation.