Update client's check for timed out key generation
As the contract is no longer enforcing timeouts for key generation we
have to change the solution where we decide how old keeps we want to
lookback for key generation. We simply refere to a time the keep was
created and define a period how long back we want to be looking.