Allow the application to set pool's minimum bondable value
BondedSortitionPool allows to specify a global minimum bondable value
the operator needs to have so that it can join and stay in the pool.
The responsibility for setting it to some reasonable value goes to the
application.

We now have two minimum bond values: one defining the minimum unbonded
value the operator needs to have so that it can join and stay in the pool,
and another defining the required bond value for the current selection.
If the given operator has enough unbonded value to stay in the pool but it
does not have enough unbonded value for the current group selection, it is
skipped.