Set the initial minimum unbonded value for a new pool to 20ETH
Sortition pool is created with a minimum bond of 20 ETH to avoid small
operators joining and griefing future selections before the minimum bond
is set to the right value by the application.
Anyone can create a sortition pool for an application with the default
minimum bond value but the application can change this value later, at
any point. 20 ETH per operator should be more than enough to cover 1 BTC
deposit in 3-operators group.