Use a more recent web3 version
We've started running into web3 errors during InitContainer runs for
both keep-client and keep-ecdsa.  Rolling the containers forward to a
newer version of web3 addresses the issue.  Along with the upgrade we've
had to adjust how we call a contract objects address, as the .address
parameter has been wrapped under a new options object.