Migrate directly from Cirlce
Despensing with the utility box since we're migrating against a public
network here.

We've had to adjust the Docker image since we need a combo of node and
gcloud to facilitate this job.  Gcloud because the
fetch-external-contracts script uses gsutil.

Additionally we requre jq but I've installed that on the job for now as
we'll probably need to roll our own image here.