Consult authorization caches before hitting ETH client
To minimize the number of calls to ETH client we introduced two caches
in the firewall: one to cache operators who do have the authorization and
another to cache operators who do not have the authorization. The
problem is that we have never added a code consulting the cache before
ETH client call is executed. Fixing it now.