Extracted random beacon consumer from keep factory
Some part of the code in ecdsa keep factory related to the random beacon
seeding could be extracted to a separate contract that is inherited by
the factory. This operation is one of the extractions from the factory
to create an abstract contract that can be used for both KEEP and ETH
staking.