Extracted submitSignatureFraud to AbstractECDSAKeep
The function verifies if the signature is fraudulent and if so it
registers the preimage and punishes members. The list of fraudulent
preimages is kept in ECDSAKeep and we can have logic there as well. The
only differenece is how we want to punish the members which we leave for
specific keep implementation as abstract function.