Fixed possibly reentrancy issue where requesting for a new seed
In case the random beacon gets compromised, it could drain factory
reseed pool to zero. The probability is very small and the pool will not
contain any significant amount but to be 100% secure, we add
nonReentrant modifier and change the order of operation a little to make
the code more clear.