Initialization period for ETH-only delegation
Added initialization period that is used to protect from attacks on
group selection. After delegation the operator has to wait the specific
period of time before being eligible to be selected to a group.

The function isInitialized is exposed by an interface required by the
FullyBackedSortitionPool for communciation with the bonding contract.

Begining of the initialiation period is the moment of delegation
receival. The `isInitialized` functions checks if the initialiation
period has passed and if the bond creator contract has been authorized
for the given operator.