Extracted current contract migration logic to a script
circleci-migrate-contracs.sh was no longer used after we decided not to
use utilitybox for migrations and all the migration logic was placed in
config.yml file. Unfortunately, the logic in config.yml missed tenderly
push command.

Here we move the migration logic to a script, and we adjust it to
install tenderly and execute tenderly push. The same logic exists in
keep-core.