Increase sortition pool join eligibility check delay to 20 minutes
Sortition pool join eligibility check was using the same retry delay as
every other retried operation in the registration logic - 1 sec. This is
pretty aggresive for this particular check as it may happen that the
client is no longer eligible to join the pool because its bonds have
been drained to zero. Here we increase the delay for this check to 20
minutes and in the future we should consider enhancing the logic in  way
allowing the operator to specify that it does no longer want to be in
the sortition pool but it still wants to do their signing duty.